Unify testing on lite container
